% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig_belonging_score.R
\name{calculate_belonging_score}
\alias{calculate_belonging_score}
\title{Calculate Belonging Score}
\usage{
calculate_belonging_score(
  motifs_scored_in_bins = motifs_scored_in_bins,
  MEAN_METHYLATION_CUTOFF = 0.25,
  N_MOTIFS_CUTOFF = 6
)
}
\arguments{
\item{motifs_scored_in_bins}{A dataframe of motifs scored in bins, containing motif
modification and bin information.}

\item{MEAN_METHYLATION_CUTOFF}{A numeric threshold for considering a motif as methylated.
Default is 0.25.}

\item{N_MOTIFS_CUTOFF}{A numeric threshold for the minimum number of motifs required for
consideration in the analysis. Default is 6.}
}
\value{
A dataframe containing the belonging scores for each bin and contig, along with
additional details such as the number of belonging bins and assembly length.
}
\description{
This function performs calculations to determine the belonging scores of motifs
based on binary comparisons of methylation patterns in contigs and bins. It includes
a scoring mechanisms considering the presence or absence of methylation
and handles special cases based on the available data.
}
\examples{
# Assuming you have a properly formatted 'motifs_scored_in_bins' dataframe:
calculate_belonging_score(motifs_scored_in_bins)

}
