% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_association.R
\name{calculate_contamination_dist_from_kmer}
\alias{calculate_contamination_dist_from_kmer}
\title{Calculate distance from Contamination to bin from K-mer Frequency}
\usage{
calculate_contamination_dist_from_kmer(
  belonging_score,
  fasta,
  contig_bins,
  kmer_window
)
}
\arguments{
\item{belonging_score}{A dataframe containing the belonging scores, used to identify potential
contamination categories.}

\item{fasta}{A list or other suitable structure containing sequences from a FASTA file,
indexed by sequence identifiers.}

\item{contig_bins}{A dataframe mapping contigs to bins.}

\item{kmer_window}{An integer specifying the k-mer window size for frequency calculation.}
}
\value{
A dataframe containing the calculated k-mer distances for each contig, which can be used
to infer contamination.
}
\description{
This function assigns contamination categories to sequences based on k-mer frequency analysis.
It involves several steps including calculation of k-mer frequencies for bins and contigs,
determination of contamination categories, and analysis of k-mer distances.
}
\examples{
# Assuming you have the necessary 'belonging_score', 'fasta', 'contig_bins', and 'kmer_window':
contamination_result <- assign_contamination_from_kmer(belonging_score, fasta, contig_bins, kmer_window)

}
