% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_association.R
\name{calculate_kmer_frequency}
\alias{calculate_kmer_frequency}
\title{Calculate K-mer Frequency}
\usage{
calculate_kmer_frequency(df, fasta, kmer_window)
}
\arguments{
\item{df}{A dataframe with a column 'contig' that corresponds to sequence identifiers in the FASTA file.}

\item{fasta}{A list or other suitable structure containing sequences from a FASTA file,
indexed by sequence identifiers.}

\item{kmer_window}{An integer specifying the k-mer window size for frequency calculation.}
}
\value{
A dataframe similar to the input, but with additional columns for each k-mer
representing its frequency count in each sequence.
}
\description{
This function calculates the k-mer frequency in sequences extracted from a FASTA file.
It processes a given dataframe to include k-mer counts for each sequence in the FASTA file
based on a specified k-mer window size.
}
\examples{
# Assuming 'df' is a dataframe and 'fasta' contains sequences:
kmer_freq_df <- calculate_kmer_frequency(df, fasta, kmer_window)

}
