% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_association.R
\name{determine_contamination}
\alias{determine_contamination}
\title{Determine Contamination Categories}
\usage{
determine_contamination(belonging_score)
}
\arguments{
\item{belonging_score}{A dataframe containing the belonging scores for each bin and contig,
as well as additional details such as the number of belonging bins.}
}
\value{
A dataframe similar to the input but with an additional column named 'group', which
indicates the contamination category for each row.
}
\description{
This function categorizes data into different contamination groups based on belonging scores.
It classifies each item as:
\itemize{
\item correct: Methylation pattern fits the assigned bin
\item single_contamination: Contig has a best match based on methylation pattern
\item multi_contamination: Contig can be assigned to multiple bins based on methylation pattern.
}
}
\examples{
# Assuming 'belonging_score' is a dataframe with the appropriate structure:
contamination_df <- determine_contamination(belonging_score)

}
