% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_association.R
\name{extract_kmer_dist}
\alias{extract_kmer_dist}
\title{Convert Distance Matrix to Long Format}
\usage{
extract_kmer_dist(dist_matrix)
}
\arguments{
\item{dist_matrix}{A distance matrix to be processed.}
}
\value{
A dataframe with two columns: 'contig' and 'matching_bins'.
The 'matching_bins' column contains concatenated distance values for each contig.
}
\description{
This function processes a distance matrix, converting it into a long format dataframe.
It isolates contig distances, rounds values, and summarizes distances by contig,
concatenating them into a single string.
}
\examples{
# Assuming 'dist_matrix' is a distance matrix:
long_format_df <- extract_kmer_dist(dist_matrix)

}
