% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig_belonging_score.R
\name{find_bins_w_no_methylation}
\alias{find_bins_w_no_methylation}
\title{Find Bins with No Methylation}
\usage{
find_bins_w_no_methylation(df, bin_motif_binary)
}
\arguments{
\item{df}{A dataframe containing methylation data, including a 'bin' column for bin identifiers
and a 'methylation_binary' column for methylation status.}

\item{bin_motif_binary}{A dataframe containing binary methylation data by bin.}
}
\value{
A dataframe either identical to the input or filtered to exclude bins with no
methylation, depending on whether such bins exist.
}
\description{
This function identifies bins with no methylation and optionally filters out these bins
from the dataset. It checks for bins where the sum of methylation binary values is zero,
indicating no methylation, and then applies this information to filter the input dataframe.
}
\examples{
# Assuming 'df' is a dataframe with methylation data and 'bin_motif_binary' contains binary methylation data:
result_df <- find_bins_w_no_methylation(df, bin_motif_binary)

}
